add DEBUG mode, in addition to ACTIVE and SLEEP.
Sometimes we need nothing other than the lua prompt, especiall to debug.
Since we are on battery all the time, we need to switch among different
modes according to one single user button, and keep the battery last
long at the same time.

Thus, to achieve all three states ACTIVE, SLEEP and DEBUG, we
designed this flow which can be described as:

       Boot
        |
        v
 button pressed? ---+
      Y |         N |
        v           |
 toggle flag_dsleep |
        |           |
        + <---------+
        |
        v
 flag_dsleep == false? ---+
      Y |               N |
        v                 v
     ACTIVE        button pressed? ---+
                        Y |         N |
                          v           v
                        DEBUG       SLEEP